Delta compression is working.
RevTree allows the caller to decide which revisions to compress when,
and relative to which references.
VersionedDoc currently has a policy of compressing all parent revisions
relative to their children, and never clearing the body (so it never
goes back to an older copy of the doc earlier in the db file anymore.)
This may not be optimal.
Additionally, it probably makes sense to compress losing conflicts
relative to their winning siblings.