Destructors should not throw exceptions
In C++11, non-throwing destructors isn't just a good idea, it's the law.
Throwing from a destructor terminates the process. Fixed several classes,
most notably Transaction -- instead of committing when the Transaction
exits scope, there's now a commit() method that has to be called. This
actually cleaned up the class significantly.

MapReduceIndexer also had to be fixed; it was committing the index
changes in its destructor.

Also marked some functions used in cleanup as `noexcept` which allows
the compiler to generate tighter cleanup code since it doesn't have to
consider the case that these functions might throw.

*Some of this should be merged into LiteCore*

For #105